Don't specify a time to "show", allow a commit instead
The @{} syntax specifies a time relative to the reflog, e.g.
master@{two.months.ago} is the commit that your "master" branch
was at two months ago, not the commits with a date just after
two months ago.  To avoid this confusion, ask for a commit
instead - people can always supply master@{two.months.ago} as
that parameter.